lms/app/requirements.txt
flask==2.3.2
flask-mysql==1.5.2
timeago==1.0.15
Flask-SQLAlchemy==3.1.1
flask-mysqldb==2.0.0
Flask_SocketIO==5.3.6